Reworked parse price logic Added parse page language

+++ Entities/Price.cs
@@ -6,7 +6,7 @@ namespace ParserFacebookShops.Entities
     {
         public string Total { get; set; }
 
-        public double? Cost { get; set; }
+        public decimal? Cost { get; set; }
 
         public string Currency { get; set; }
